# common-lib/src/main/java/com/lumastyle/common/error/ErrorResponse.java
package com.lumastyle.common.error;

import lombok.Builder;
import lombok.Getter;

import java.time.Instant;
import java.util.Map;

@Getter
@Builder
public class ErrorResponse {
    private Instant timestamp;
    private int status;
    private String error;
    private String message;
    private Map<String, String> errors; // Optional field errors
}